+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/block/BlockConstants.java
@@ -2,16 +2,15 @@ package com.refinedmods.refinedstorage2.platform.common.block;
 
 import net.minecraft.world.level.block.SoundType;
 import net.minecraft.world.level.block.state.BlockBehaviour;
-import net.minecraft.world.level.material.Material;
 
 public final class BlockConstants {
     public static final BlockBehaviour.Properties PROPERTIES = BlockBehaviour.Properties
-        .of(Material.STONE)
+        .of()
         .strength(0.5F, 6.0F)
         .sound(SoundType.STONE);
 
     public static final BlockBehaviour.Properties CABLE_PROPERTIES = BlockBehaviour.Properties
-        .of(Material.GLASS)
+        .of()
         .strength(0.35F, 0.35F)
         .sound(SoundType.GLASS);