+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/screen/ExternalStorageScreen.java
@@ -33,27 +33,14 @@ public class ExternalStorageScreen extends AbstractBaseScreen<ExternalStorageCon
     protected void init() {
         super.init();
 
-        addSideButton(new RedstoneModeSideButtonWidget(
-            getMenu().getProperty(PropertyTypes.REDSTONE_MODE),
-            this::renderComponentTooltip
-        ));
-        addSideButton(new FilterModeSideButtonWidget(
-            getMenu().getProperty(PropertyTypes.FILTER_MODE),
-            this::renderComponentTooltip
-        ));
-        addSideButton(new FuzzyModeSideButtonWidget(
-            getMenu().getProperty(PropertyTypes.FUZZY_MODE),
-            this::renderComponentTooltip
-        ));
-        addSideButton(new AccessModeSideButtonWidget(
-            getMenu().getProperty(PropertyTypes.ACCESS_MODE),
-            this::renderComponentTooltip
-        ));
+        addSideButton(new RedstoneModeSideButtonWidget(getMenu().getProperty(PropertyTypes.REDSTONE_MODE)));
+        addSideButton(new FilterModeSideButtonWidget(getMenu().getProperty(PropertyTypes.FILTER_MODE)));
+        addSideButton(new FuzzyModeSideButtonWidget(getMenu().getProperty(PropertyTypes.FUZZY_MODE)));
+        addSideButton(new AccessModeSideButtonWidget(getMenu().getProperty(PropertyTypes.ACCESS_MODE)));
         addSideButton(new PrioritySideButtonWidget(
             getMenu().getProperty(PropertyTypes.PRIORITY),
             playerInventory,
-            this,
-            this::renderComponentTooltip
+            this
         ));
     }