+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/screen/TooltipRenderer.java
@@ -2,10 +2,10 @@ package com.refinedmods.refinedstorage2.platform.common.screen;
 
 import java.util.List;
 
-import com.mojang.blaze3d.vertex.PoseStack;
+import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.network.chat.Component;
 
 @FunctionalInterface
 public interface TooltipRenderer {
-    void render(PoseStack poseStack, List<Component> lines, int x, int y);
+    void render(GuiGraphics graphics, List<Component> lines, int x, int y);
 }