+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/screen/grid/SortingTypeSideButtonWidget.java
@@ -2,7 +2,6 @@ package com.refinedmods.refinedstorage2.platform.common.screen.grid;
 
 import com.refinedmods.refinedstorage2.platform.common.containermenu.grid.AbstractGridContainerMenu;
 import com.refinedmods.refinedstorage2.platform.common.internal.grid.GridSortingTypes;
-import com.refinedmods.refinedstorage2.platform.common.screen.TooltipRenderer;
 import com.refinedmods.refinedstorage2.platform.common.screen.widget.AbstractSideButtonWidget;
 
 import java.util.ArrayList;
@@ -12,7 +11,6 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 
-import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
 
@@ -20,13 +18,11 @@ import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUti
 
 public class SortingTypeSideButtonWidget extends AbstractSideButtonWidget {
     private final AbstractGridContainerMenu menu;
-    private final TooltipRenderer tooltipRenderer;
     private final Map<GridSortingTypes, List<Component>> tooltips = new EnumMap<>(GridSortingTypes.class);
 
-    public SortingTypeSideButtonWidget(final AbstractGridContainerMenu menu, final TooltipRenderer tooltipRenderer) {
+    public SortingTypeSideButtonWidget(final AbstractGridContainerMenu menu) {
         super(createPressAction(menu));
         this.menu = menu;
-        this.tooltipRenderer = tooltipRenderer;
         Arrays.stream(GridSortingTypes.values()).forEach(type -> tooltips.put(type, calculateTooltip(type)));
     }
 
@@ -69,7 +65,7 @@ public class SortingTypeSideButtonWidget extends AbstractSideButtonWidget {
     }
 
     @Override
-    public void onTooltip(final PoseStack poseStack, final int mouseX, final int mouseY) {
-        tooltipRenderer.render(poseStack, tooltips.get(menu.getSortingType()), mouseX, mouseY);
+    protected List<Component> getSideButtonTooltip() {
+        return tooltips.get(menu.getSortingType());
     }
 }