+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/render/model/baked/transform/EmissiveTransform.java
@@ -27,7 +27,7 @@ public class EmissiveTransform implements RenderContext.QuadTransform {
         if (finder == null) {
             return;
         }
-        final TextureAtlasSprite sprite = finder.find(quad, 0);
+        final TextureAtlasSprite sprite = finder.find(quad);
         if (sprite == null) {
             return;
         }